fix: fix movie budget and revenue integer overflow error
Change type of budget and revenue to Long

+++ app/src/main/java/io/woong/filmpedia/ui/page/moviedetail/MovieDetailActivity.kt
@@ -305,7 +305,7 @@ fun AppCompatTextView.bindProductionCompanies(companies: List<Movie.Company>?) {
 }
 
 @BindingAdapter("movie_detail_money")
-fun AppCompatTextView.bindMoney(money: Int?) {
+fun AppCompatTextView.bindMoney(money: Long?) {
     money?.let { m ->
         if (m <= 0) {
             this.text = "-"
